[proj/stm32] add 5 processes in total for test.
Finally we made it!

We have nothing but a rough enough schduler which can perform muti-process.
No ipc, no lock, no interrupt handling mechanism...
The only things we can do is to yield, to sleep, and to schedule!
I believe that's all for a basis of the ugliest/simplest OS.

Let's upgrade to 0.1.0 :P