plat: stm32: fix usb console issues
It would block until we open the host side cdc/acm device (i.e. use
minicom), because the systick is not enabled yet when we print the
first line to usb console. And the root cause is the variable -
"systick_uptime_millis" get from systick_uptime() was always 0 and
the USB_TIMEOUT was no use.

Now we correct that thing and it will enable a USB_TIMEOUT
millisecond busywait while doing usb printing.
Thus we will never *block* anymore although the performance will be
worse while enable the usb console.

Anyway, things are not always bad or good. It is easier for debug,
with a worse performance. Now set a 2 milliseconds timeout should
be accepted.

Summary:
1. Fix systick_uptime_millis always being 0 which would cause block.
2. Enable usb printing things only after systick enabled.